# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "My first Go API",
        "title": "My Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
        "/empinfo": {
            "get": {
                "description": "Returns employees with department and bio info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get employee info",
                "parameters": [
                    {
                        "type": "string",
                        "default": "",
                        "description": "Department name to filter by",
                        "name": "Dept",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "Filter by Emp Name to filter by",
                        "name": "FilterName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.EmpInfo"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Returns a list of employees",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get employees",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Employee"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.EmpInfo": {
            "type": "object",
            "properties": {
                "BioID": {
                    "type": "string"
                },
                "Dept": {
                    "type": "string"
                },
                "IdNo": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "main.Employee": {
            "type": "object",
            "properties": {
                "fname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lname": {
                    "type": "string"
                }
            }
        }
    }
}